first_langchain/pyproject.toml
[tool.poetry]
name = "first-langchain"
version = "0.1.0"
description = ""
authors = ["Evan <evanmschultz@gmail.com>"]
readme = "README.md"
packages = [{include = "first_langchain"}]

[tool.poetry.dependencies]
python = "^3.11"
langchain = "^0.0.262"
openai = "^0.27.8"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
